packages/observability/pyproject.toml
[project]
name = "arachne-observability"
version = "0.1.0"
description = "Structured logging and OpenTelemetry instrumentation for Arachne"
requires-python = ">=3.13"
dependencies = [
    "structlog>=24.4,<25",
    "opentelemetry-api>=1.29,<2",
    "opentelemetry-sdk>=1.29,<2",
    "opentelemetry-exporter-otlp-proto-grpc>=1.29,<2",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "opentelemetry-instrumentation-httpx>=0.50b0",
    "prometheus-client>=0.21,<1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/arachne_observability"]

[tool.ruff]
target-version = "py313"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]
